refactor(events): extract shared event-form helpers for Wizard/Editor (#544)
EventWizard.svelte (871) and EventEditor.svelte (775) carried three sets of
duplicated logic verbatim. Hoist the identical copies into co-located modules:

- event-associations.ts: saveResourceAssociations / saveTagAssociations
  (identical diff-and-sync logic; tag helper returns the new initial-tags
  snapshot so the reassignment side effect stays in the component).
- event-image-mutations.svelte.ts: createEventImageMutations() factory for the
  four upload/delete logo & cover-art mutations (pure mutationFns, no component
  state — byte-for-byte identical between the two files).
- event-payload.ts: buildEventCreateData (shared, identical) plus the divergent
  update builders kept as honestly-separate functions — buildWizardStep1/Step2
  UpdateData and buildEditorUpdateData (the latter alone carries
  public_pronoun_distribution). is_open_ended end-null handling preserved.

EventWizard 871 -> 660, EventEditor 775 -> 646. Behavior unchanged.
